requirements.txt
mysql-connector-python==8.1.0
protobuf==4.21.12
tqdm==4.65.0
